+++ apps/rahat-ui/src/app/projects/layout.tsx
@@ -4,6 +4,7 @@ import { useParams, usePathname, useRouter } from 'next/navigation';
 import * as React from 'react';
 import { ProjectLayout } from '../../sections/projects/components';
 import DashboardLayout from '../dashboard/layout';
+import AuthGuard from '../../guards/auth-guard';
 
 export default function ProjectLayoutRoot({
   children,
@@ -14,32 +15,20 @@ export default function ProjectLayoutRoot({
   const allowedPaths: string[] = [];
   const allowNavPaths = ['/projects'];
   // const allowedPaths = ['/projects', '/projects/add'];
-  //
-  const { id } = useParams();
-  const router = useRouter();
 
-  // UUID format validation (simple regex for UUID v4)
-  const isValidUUID = (uuid: string) =>
-    /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$/.test(
-      uuid,
-    );
-  React.useEffect(() => {
-    if (!id || !isValidUUID(id)) {
-      // Redirect to the project list page if UUID is missing or invalid
-      router.push('/projects');
-    }
-  }, [id]);
   return (
-    <DashboardLayout
-      hasDefaultHeader={allowNavPaths.includes(pathName)}
-      margin="mt-0"
-    >
-      <title>Projects</title>
-      {!allowedPaths.includes(pathName) ? (
-        <>{children}</>
-      ) : (
-        <ProjectLayout projectType="ALL">{children}</ProjectLayout>
-      )}
-    </DashboardLayout>
+    <AuthGuard>
+      <DashboardLayout
+        hasDefaultHeader={allowNavPaths.includes(pathName)}
+        margin="mt-0"
+      >
+        <title>Projects</title>
+        {!allowedPaths.includes(pathName) ? (
+          <>{children}</>
+        ) : (
+          <ProjectLayout projectType="ALL">{children}</ProjectLayout>
+        )}
+      </DashboardLayout>
+    </AuthGuard>
   );
 }
