+++ apps/rahat-ui/src/sections/projects/components/project.layout.tsx
@@ -5,7 +5,7 @@ import {
   ResizablePanel,
   ResizablePanelGroup,
 } from '@rahat-ui/shadcn/src/components/ui/resizable';
-import { FC } from 'react';
+import React, { FC } from 'react';
 import { ProjectType } from './nav-items.types';
 import { ProjectNav } from './project-header';
 import ProjectNavView from './project.nav.view';
@@ -13,6 +13,7 @@ import { useProjectHeaderItems } from './useProjectHeaderItems';
 import { useProjectNavItems } from './useProjectNavItems';
 import { SidebarProvider } from '@rahat-ui/shadcn/src/components/ui/sidebar';
 import { ProjectSidebar } from 'apps/rahat-ui/src/sidebar-components/project-sidebar';
+import { useParams, useRouter } from 'next/navigation';
 
 type ProjectLayoutProps = {
   children: React.ReactNode | React.ReactNode[];
@@ -25,6 +26,21 @@ const ProjectLayout: FC<ProjectLayoutProps> = ({
   projectType,
   navFooter,
 }) => {
+  const { id } = useParams();
+  const router = useRouter();
+
+  // UUID format validation (simple regex for UUID v4)
+  const isValidUUID = (uuid: string) =>
+    /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$/.test(
+      uuid,
+    );
+  React.useEffect(() => {
+    if (!id || !isValidUUID(id)) {
+      // Redirect to the project list page if UUID is missing or invalid
+      router.push('/projects');
+    }
+  }, [id]);
+
   const { navItems: menuItems } = useProjectNavItems(projectType);
   const { headerNav } = useProjectHeaderItems(projectType);
   const renderResizablePanel = (children: React.ReactNode, index?: number) => {
