+++ apps/rahat-ui/src/sections/projects/el-kenya/vouchers/vouchers.manage.tsx
@@ -16,6 +16,7 @@ import {
   useReadRahatTokenTotalSupply,
 } from '@rahat-ui/query';
 import { UUID } from 'crypto';
+import { useAccount } from 'wagmi';
 
 interface VouchersManageProps {
   handleStepDataChange: (e) => void;
@@ -31,6 +32,7 @@ export default function VouchersManage({
   stepData,
 }: VouchersManageProps) {
   const { id } = useParams() as { id: UUID };
+  const { isConnected } = useAccount();
 
   const contractSettings = useProjectSettingsStore(
     (state) => state.settings?.[id]?.[PROJECT_SETTINGS_KEYS.CONTRACT],
@@ -52,54 +54,62 @@ export default function VouchersManage({
         Icon={Ticket}
         number={tokenBalance?.toString() ?? '-'}
       />
-      <Tabs defaultValue="beneficiary">
-        <TabsContent value="beneficiary">
-          <div className="pl-4">
-            <h1 className=" font-semibold text-xl ">Beneficiary</h1>
-            <p className="text-muted-foreground text-base">
-              List of all the beneficiaries assigned to the vendors
-            </p>
-          </div>
-        </TabsContent>
-        <TabsContent value="beneficiaryGroups">
-          <div className="pl-4">
-            <h1 className=" font-semibold text-xl ">Beneficiary Groups</h1>
-            <p className="text-muted-foreground text-base">
-              List of all the beneficiaries groups assigned to the vendors
-            </p>
-          </div>
-        </TabsContent>
-        <div className="flex justify-between items-center p-4 pt-3 pb-0">
-          <TabsList className="border bg-secondary rounded">
-            <TabsTrigger
-              className="w-full data-[state=active]:bg-white"
-              value="beneficiary"
-            >
-              Beneficiary
-            </TabsTrigger>
-            <TabsTrigger
-              className="w-full data-[state=active]:bg-white"
-              value="beneficiaryGroups"
-            >
-              Beneficiary Groups
-            </TabsTrigger>
-          </TabsList>
+      {!isConnected ? (
+        <div className=" flex justify-center items-center mb-4">
+          <p className="text-muted-foreground">
+            Please connect your wallet to proceed
+          </p>
         </div>
-        <TabsContent value="beneficiary">
-          <BeneficiaryView
-            handleStepDataChange={handleStepDataChange}
-            handleNext={handleNext}
-          />
-        </TabsContent>
-        <TabsContent value="beneficiaryGroups">
-          <BeneficiaryGroupsView
-            handleStepDataChange={handleStepDataChange}
-            handleNext={handleNext}
-            setBeneficiaryGroupSelected={setBeneficiaryGroupSelected}
-            stepData={stepData}
-          />
-        </TabsContent>
-      </Tabs>
+      ) : (
+        <Tabs defaultValue="beneficiary">
+          <TabsContent value="beneficiary">
+            <div className="pl-4">
+              <h1 className=" font-semibold text-xl ">Beneficiary</h1>
+              <p className="text-muted-foreground text-base">
+                List of all the beneficiaries assigned to the vendors
+              </p>
+            </div>
+          </TabsContent>
+          <TabsContent value="beneficiaryGroups">
+            <div className="pl-4">
+              <h1 className=" font-semibold text-xl ">Beneficiary Groups</h1>
+              <p className="text-muted-foreground text-base">
+                List of all the beneficiaries groups assigned to the vendors
+              </p>
+            </div>
+          </TabsContent>
+          <div className="flex justify-between items-center p-4 pt-3 pb-0">
+            <TabsList className="border bg-secondary rounded">
+              <TabsTrigger
+                className="w-full data-[state=active]:bg-white"
+                value="beneficiary"
+              >
+                Beneficiary
+              </TabsTrigger>
+              <TabsTrigger
+                className="w-full data-[state=active]:bg-white"
+                value="beneficiaryGroups"
+              >
+                Beneficiary Groups
+              </TabsTrigger>
+            </TabsList>
+          </div>
+          <TabsContent value="beneficiary">
+            <BeneficiaryView
+              handleStepDataChange={handleStepDataChange}
+              handleNext={handleNext}
+            />
+          </TabsContent>
+          <TabsContent value="beneficiaryGroups">
+            <BeneficiaryGroupsView
+              handleStepDataChange={handleStepDataChange}
+              handleNext={handleNext}
+              setBeneficiaryGroupSelected={setBeneficiaryGroupSelected}
+              stepData={stepData}
+            />
+          </TabsContent>
+        </Tabs>
+      )}
     </div>
   );
 }
